refactor(tests): move unique_filepath test to test_system_utils
Create tests/utils_/test_system_utils.py and move test_unique_filepath
from test_utils.py to match the code refactoring where unique_filepath
was moved to vllm.utils.system_utils module.
